pyproject.toml
[project]
name = "backend-estoque"
version = "0.1.0"
description = ""
authors = [
    {name = "T-Terra",email = "gabst138@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
mkdocs = "^1.6.1"
pip-audit = "^2.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
